prepare using url overrides
