+++ src/buildProject.js
@@ -29,6 +29,7 @@ export const buildProject = async ({
   globalName,
   globals = {},
 
+  urlOverrides = {},
   importResolutionMethod = format === "commonjs" ? "node" : "importmap",
   importMapFileRelativeUrl,
   importDefaultExtension,
@@ -165,6 +166,7 @@ export const buildProject = async ({
         compileServerOrigin,
         compileDirectoryRelativeUrl: `${outDirectoryRelativeUrl}${COMPILE_ID_OTHERWISE}/`,
 
+        urlOverrides,
         importResolutionMethod,
         importMapFileRelativeUrl,
         importDefaultExtension,
